+++ email-worker/src/email/constant.ts
@@ -1,0 +1,7 @@
+export const EmailPayloadConstant = {
+  USER_CERTIFICATION: 'userCertification',
+  RSS_REMOVAL: 'rssRemoval',
+  RSS_REGISTRATION: 'rssRegistration',
+  PASSWORD_RESET: 'passwordReset',
+  ACCOUNT_DELETION: 'accountDeletion',
+} as const;
