+++ email-worker/src/types/types.ts
@@ -24,20 +24,3 @@ export interface RssRemoval {
   certificateCode: string;
 }
 
-export const EmailPayloadConstant = {
-  USER_CERTIFICATION: 'userCertification',
-  RSS_REMOVAL: 'rssRemoval',
-  RSS_REGISTRATION: 'rssRegistration',
-  PASSWORD_RESET: 'passwordReset',
-  ACCOUNT_DELETION: 'accountDeletion',
-} as const;
-
-export type EmailPayload =
-  | { type: typeof EmailPayloadConstant.USER_CERTIFICATION; data: User }
-  | { type: typeof EmailPayloadConstant.RSS_REMOVAL; data: RssRemoval }
-  | {
-      type: typeof EmailPayloadConstant.RSS_REGISTRATION;
-      data: RssRegistration;
-    }
-  | { type: typeof EmailPayloadConstant.PASSWORD_RESET; data: User }
-  | { type: typeof EmailPayloadConstant.ACCOUNT_DELETION; data: User };
