+++ email-worker/test/unit/consumer.spec.ts
@@ -1,7 +1,9 @@
 import 'reflect-metadata';
 
+import { EmailPayloadConstant } from '@email/constant';
 import { EmailConsumer } from '@email/email.consumer';
 import { EmailService } from '@email/email.service';
+import { EmailPayload } from '@email/type';
 
 import { Notifier } from '@notification/notifier.interface';
 
@@ -9,13 +11,7 @@ import { RETRY_CONFIG, RMQ_QUEUES } from '@rabbitmq/rabbitmq.constant';
 import { RabbitMQService } from '@rabbitmq/rabbitmq.service';
 
 import { NodeMailerError } from '@app-types/nodemailer';
-import {
-  EmailPayload,
-  EmailPayloadConstant,
-  RssRegistration,
-  RssRemoval,
-  User,
-} from '@app-types/types';
+import { RssRegistration, RssRemoval, User } from '@app-types/types';
 
 describe('email consumer unit test', () => {
   let emailConsumer: EmailConsumer;
